Agrega paginacion a las tablas para administrar. Reorganiza los componentes y agrega paginacion.
