+++ client/src/components/PcBuildTables/MotherboardTable.tsx
@@ -1,0 +1,156 @@
+import { Box, Button, ButtonGroup } from "@material-ui/core";
+import {
+  DataGrid,
+  GridCellParams,
+  GridColDef,
+  GridPageChangeParams,
+  ValueFormatterParams,
+} from "@material-ui/data-grid";
+import AddCircleOutlineIcon from "@material-ui/icons/AddCircleOutline";
+import { useState } from "react";
+import { Link as RouterLink } from "react-router-dom";
+import { IGetMotherboards } from "src/hooks/types";
+import useDelete from "src/hooks/useDelete";
+import useGetAll from "src/hooks/useGetAll";
+import { IMotherboard } from "src/types";
+import currencyFormatter from "src/utils/formatCurrency";
+import CustomToolbar from "../CustomToolbar";
+import DeleteDialog from "../DeleteDialogs/DeleteDialog";
+import SnackbarAlert from "../SnackbarAlert";
+
+const MotherboardContainer = () => {
+  const PAGE_SIZE = 12;
+  const [offset, setOffset] = useState(0);
+  const [open, setOpen] = useState(false);
+  const [deleteId, setDeleteId] = useState<string>("");
+  const deleteMotherboard = useDelete<IMotherboard>("motherboard");
+  const queryMotherboards = useGetAll<IGetMotherboards>("motherboard", offset, PAGE_SIZE);
+
+  const handlePageChange = (params: GridPageChangeParams) => {
+    setOffset(params.page * PAGE_SIZE);
+  };
+
+  const handleClickDeleteBtn = (id: string) => {
+    setOpen(true);
+    setDeleteId(id);
+  };
+
+  const closeDialog = () => {
+    setOpen(false);
+  };
+
+  const handleDelete = () => {
+    deleteMotherboard.mutate(deleteId);
+    closeDialog();
+  };
+
+  return (
+    <>
+      {deleteMotherboard.isSuccess ? (
+        <SnackbarAlert severity="success" text="Motherboard deleted successfully"></SnackbarAlert>
+      ) : (
+        deleteMotherboard.isError && <SnackbarAlert severity="error" text="Something went wrong"></SnackbarAlert>
+      )}
+
+      <Box mb={1}>
+        <Button to="create/motherboard" component={RouterLink} variant="outlined" endIcon={<AddCircleOutlineIcon />}>
+          Add new motherboard
+        </Button>
+      </Box>
+
+      <DeleteDialog toDelete="motherboard" open={open} closeDialog={closeDialog} handleDelete={handleDelete} />
+
+      <Box height="525px" marginBottom="50px">
+        <DataGrid
+          pagination
+          paginationMode="server"
+          pageSize={PAGE_SIZE}
+          rowCount={queryMotherboards.isSuccess ? queryMotherboards.data.count : undefined}
+          onPageChange={handlePageChange}
+          loading={queryMotherboards.isLoading}
+          columns={
+            [
+              { field: "id", type: "number", hide: true },
+              { field: "product_id", type: "number", hide: true },
+              { field: "name", width: 200, headerName: "Name" },
+              { field: "description", width: 200, headerName: "Description" },
+              {
+                field: "price",
+                width: 140,
+                headerName: "Price",
+                headerAlign: "left",
+                align: "center",
+                type: "number",
+                valueFormatter: (params: ValueFormatterParams) => currencyFormatter.format(Number(params.value)),
+              },
+              { field: "stock", width: 100, headerName: "Stock" },
+              { field: "brand", width: 100, headerName: "Brand" },
+              { field: "cpu_brand", width: 150, headerName: "CPU Brand" },
+              { field: "cpu_socket", width: 150, headerName: "CPU Socket" },
+              {
+                field: "min_frec",
+                width: 120,
+                headerName: "Min Frec",
+                type: "number",
+                valueFormatter: (params: ValueFormatterParams) => params.value + " MHz",
+              },
+              {
+                field: "max_frec",
+                width: 120,
+                headerName: "Max Frec",
+                type: "number",
+                valueFormatter: (params: ValueFormatterParams) => params.value + " MHz",
+              },
+              { field: "ram_version", width: 150, headerName: "RAM version" },
+              { field: "model_size", width: 150, headerName: "Model Size" },
+              { field: "video_socket", width: 150, headerName: "Video Socket" },
+              {
+                field: "edit",
+                headerName: "Edit options",
+                sortable: false,
+                filterable: false,
+                width: 450,
+                renderCell: (params: GridCellParams) => (
+                  <ButtonGroup>
+                    <Button to={"/admin/products/edit/" + params.row.product_id} component={RouterLink}>
+                      Edit product
+                    </Button>
+                    <Button to={"edit/motherboard/" + params.row.id} component={RouterLink}>
+                      Edit specs
+                    </Button>
+                    <Button onClick={() => handleClickDeleteBtn(params.row.id as string)}>Delete</Button>
+                  </ButtonGroup>
+                ),
+              },
+            ] as GridColDef[]
+          }
+          rows={
+            queryMotherboards.isSuccess
+              ? queryMotherboards.data.results.map((motherboard: IMotherboard) => ({
+                  id: motherboard.id,
+                  product_id: motherboard.product?.id,
+                  name: motherboard.product?.name,
+                  description: motherboard.product?.description,
+                  stock: motherboard.product?.stock ? "Yes" : "No",
+                  price: motherboard.product!.price,
+                  brand: motherboard.product?.brand.name || "Not found",
+                  cpu_brand: motherboard.cpu_brand,
+                  cpu_socket: motherboard.cpu_socket,
+                  min_frec: motherboard.min_frec,
+                  max_frec: motherboard.max_frec,
+                  model_size: motherboard.model_size,
+                  ram_version: motherboard.ram_version,
+                  video_socket: motherboard.video_socket,
+                }))
+              : []
+          }
+          components={{
+            Toolbar: CustomToolbar,
+          }}
+        />
+      </Box>
+    </>
+  );
+};
+
+export default MotherboardContainer;
